Refactor: Button컴포넌트사용해서 버튼 수정, SelectBox.tsx atom파일로 이동 #106
